Batch execution of JIT SLT tests
