+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/OutputViewDescription.java
@@ -7,6 +7,10 @@ import org.dbsp.sqlCompiler.compiler.backend.jit.JitJsonOutputDescription;
 import org.dbsp.sqlCompiler.compiler.errors.UnimplementedException;
 import org.dbsp.sqlCompiler.compiler.frontend.calciteCompiler.RelColumnMetadata;
 import org.dbsp.sqlCompiler.compiler.frontend.statements.CreateViewStatement;
+import org.dbsp.util.FreshName;
+
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * Information about an output view.
@@ -22,8 +26,14 @@ public class OutputViewDescription {
         switch (fas.kind) {
             case Json: {
                 JitJsonOutputDescription result = new JitJsonOutputDescription(this.getName(), fas.path);
-                for (RelColumnMetadata column: this.view.columns)
-                    result.addColumn(column.getName(), column.getType());
+                FreshName fresh = new FreshName(new HashSet<>());
+                for (RelColumnMetadata column: this.view.columns) {
+                    String columnName = fresh.freshName(column.getName());
+                    // May not be exactly like the original column name if we have
+                    // duplicates, but no one really looks at these, and this
+                    // code is only used for testing.
+                    result.addColumn(columnName, column.getType());
+                }
                 return result;
             }
             default:
