+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqllogictest/executors/JitDbspExecutor.java
@@ -1,49 +1,21 @@
 package org.dbsp.sqllogictest.executors;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import net.hydromatic.sqllogictest.ISqlTestOperation;
-import net.hydromatic.sqllogictest.OptionsParser;
-import net.hydromatic.sqllogictest.SltSqlStatement;
-import net.hydromatic.sqllogictest.SltTestFile;
-import net.hydromatic.sqllogictest.SqlTestQuery;
-import net.hydromatic.sqllogictest.SqlTestQueryOutputDescription;
-import net.hydromatic.sqllogictest.TestStatistics;
+import net.hydromatic.sqllogictest.*;
 import net.hydromatic.sqllogictest.executors.JdbcExecutor;
 import net.hydromatic.sqllogictest.executors.SqlSltTestExecutor;
-import org.dbsp.sqlCompiler.circuit.DBSPCircuit;
 import org.dbsp.sqlCompiler.compiler.CompilerOptions;
-import org.dbsp.sqlCompiler.compiler.DBSPCompiler;
-import org.dbsp.sqlCompiler.compiler.backend.ToCsvVisitor;
-import org.dbsp.sqlCompiler.compiler.backend.jit.JitFileAndSerialization;
-import org.dbsp.sqlCompiler.compiler.backend.jit.JitIODescription;
-import org.dbsp.sqlCompiler.compiler.backend.jit.JitSerializationKind;
-import org.dbsp.sqlCompiler.compiler.backend.jit.ToJitVisitor;
-import org.dbsp.sqlCompiler.compiler.backend.jit.ir.JITProgram;
-import org.dbsp.sqlCompiler.ir.expression.DBSPExpression;
-import org.dbsp.sqlCompiler.ir.expression.DBSPTupleExpression;
-import org.dbsp.sqlCompiler.ir.expression.literal.*;
-import org.dbsp.sqlCompiler.ir.type.DBSPType;
-import org.dbsp.sqlCompiler.ir.type.DBSPTypeZSet;
+import org.dbsp.sqlCompiler.ir.expression.literal.DBSPZSetLiteral;
 import org.dbsp.sqllogictest.SqlTestPrepareInput;
 import org.dbsp.sqllogictest.SqlTestPrepareTables;
 import org.dbsp.sqllogictest.SqlTestPrepareViews;
-import org.dbsp.util.Linq;
-import org.dbsp.util.Utilities;
 
 import javax.annotation.Nullable;
-import java.io.File;
 import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Paths;
-import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 import java.util.regex.Matcher;
@@ -59,7 +31,7 @@ public class JitDbspExecutor extends SqlSltTestExecutor {
     final SqlTestPrepareViews viewPreparation;
     protected final CompilerOptions compilerOptions;
     static final String rustDirectory = "../temp/src/";
-    static final String projectDirectory = "..";
+    static final String jitExecutableDirectory = "..";
 
     public JitDbspExecutor(JdbcExecutor executor,
                            OptionsParser.SuppliedOptions options,
@@ -79,13 +51,6 @@ public class JitDbspExecutor extends SqlSltTestExecutor {
 
     DBSPZSetLiteral getTableContents(String table) throws SQLException {
         return DbspJdbcExecutor.getTableContents(this.getStatementExecutorConnection(), table);
-    }
-
-    void createTables(DBSPCompiler compiler) {
-        for (SltSqlStatement statement : this.tablePreparation.statements) {
-            String stat = statement.statement;
-            compiler.compileStatement(stat, stat);
-        }
     }
 
     @Nullable
@@ -137,381 +102,21 @@ public class JitDbspExecutor extends SqlSltTestExecutor {
         return result;
     }
 
-    File fileFromName(String name) {
-        return new File(Paths.get(rustDirectory, name).toUri());
-    }
-
-    static DBSPExecutor.TableValue[] previousValues = new DBSPExecutor.TableValue[0];
-
-    boolean query(SqlTestQuery query, TestStatistics statistics, int queryNo)
-            throws IOException, InterruptedException, NoSuchAlgorithmException, SQLException {
-        if (this.buggyOperations.contains(query.getQuery())
-                || this.options.doNotExecute) {
-            statistics.incIgnored();
-            options.message("Skipping " + query.getQuery(), 2);
-            return false;
-        }
-
-        DBSPCompiler compiler = new DBSPCompiler(this.compilerOptions);
-        this.createTables(compiler);
-        String dbspQuery = query.getQuery();
-        if (!dbspQuery.toLowerCase().contains("create view"))
-            dbspQuery = "CREATE VIEW V AS (" + dbspQuery + ")";
-        this.options.message("Query " + queryNo + ":\n"
-                + dbspQuery + "\n", 2);
-        compiler.generateOutputForNextView(false);
-        for (SltSqlStatement view: viewPreparation.definitions()) {
-            compiler.compileStatement(view.statement, view.statement);
-            compiler.throwIfErrorsOccurred();
-        }
-        compiler.generateOutputForNextView(true);
-        compiler.compileStatement(dbspQuery, "" /* testQuery.getName() */);
-        compiler.throwIfErrorsOccurred();
-        DBSPExecutor.TableValue[] inputSets = this.getInputSets();
-
-        compiler.optimize();
-        DBSPCircuit circuit = compiler.getFinalCircuit("circuit");
-        // Serialize circuit as JSON for the JIT executor
-        JITProgram program = ToJitVisitor.circuitToJIT(compiler, circuit);
-        String json = program.asJson().toPrettyString();
-
-        List<File> toDelete = new ArrayList<>();
-        File programFile = this.fileFromName("program.json");
-        toDelete.add(programFile);
-        Utilities.writeFile(programFile.toPath(), json);
-        File asm = this.fileFromName("program.asm");
-        toDelete.add(asm);
-        Utilities.writeFile(asm.toPath(), program.toString());
-
-        // Prepare input files for the JIT runtime
-        boolean sameFiles = Arrays.equals(previousValues, inputSets);
-        List<JitFileAndSerialization> inputFiles = new ArrayList<>();
-        int index = 0;
-        for (DBSPExecutor.TableValue inputData : inputSets) {
-            File input = this.fileFromName("input" + index++ + ".csv");
-            if (!sameFiles) {
-                ToCsvVisitor.toCsv(compiler, input, new DBSPZSetLiteral(
-                        compiler.getWeightTypeImplementation(), inputData.contents.data));
-            }
-            inputFiles.add(new JitFileAndSerialization(
-                    input.getAbsolutePath(),
-                    JitSerializationKind.Csv));
-        }
-        previousValues = inputSets;
-        List<JitIODescription> inputDescriptions = compiler.getInputDescriptions(inputFiles);
-
-        // Allocate output files
-        List<JitFileAndSerialization> outputFiles = new ArrayList<>();
-        if (circuit.getOutputCount() != 1)
-            throw new RuntimeException("Expected a single output");
-        File output = this.fileFromName("output.json");
-        toDelete.add(output);
-        outputFiles.add(new JitFileAndSerialization(
-                output.getAbsolutePath(),
-                JitSerializationKind.Json));
-        List<JitIODescription> outputDescriptions = compiler.getOutputDescriptions(outputFiles);
-
-        // Invoke the JIT runtime with the program and the configuration file describing inputs and outputs
-        JsonNode jitInputDescription = compiler.createJitRuntimeConfig(inputDescriptions, outputDescriptions);
-        String s = jitInputDescription.toPrettyString();
-        File configFile = this.fileFromName("config.json");
-        toDelete.add(configFile);
-        Utilities.writeFile(configFile.toPath(), s);
-        Utilities.runJIT(projectDirectory, programFile.getAbsolutePath(), configFile.getAbsolutePath());
-
-        // Validate outputs and delete them
-        DBSPType outputType = circuit.getOutputType(0);
-        JitIODescription outFile = outputDescriptions.get(0);
-        DBSPZSetLiteral.Contents actual = outFile.parse(outputType.to(DBSPTypeZSet.class).getElementType());
-
-        boolean result = this.validateOutput(query, queryNo, actual, query.outputDescription, statistics);
-        if (!result) {
-            for (File file : toDelete) {
-                // This point won't be reached if the program fails with an exception
-                // or validation fails with -x flag
-                boolean success = file.delete();
-                if (!success)
-                    System.err.println("Failed to delete " + file);
-            }
-        }
-        return result;
-    }
-
     // These should be reused from JdbcExecutor, but are not public there.
-    /**
-     * A row produced by a query execution.
-     */
-    static class Row {
-        /**
-         * In SLT all data received from the database is converted to strings.
-         */
-        public final List<String> values;
-        Row() {
-            this.values = new ArrayList<>();
-        }
-        void add(String v) {
-            this.values.add(v);
-        }
-        @Override public String toString() {
-            return String.join(System.lineSeparator(), this.values);
-        }
-    }
-
-    static class RowComparator implements Comparator<Row> {
-        @Override public int compare(Row o1, Row o2) {
-            if (o1.values.size() != o2.values.size()) {
-                throw new RuntimeException("Comparing rows of different lengths");
-            }
-            for (int i = 0; i < o1.values.size(); i++) {
-                int r = o1.values.get(i).compareTo(o2.values.get(i));
-                if (r != 0) {
-                    return r;
-                }
-            }
-            return 0;
-        }
-    }
 
     /**
-     * A set of rows produced as a result of a query execution.
+     * Run a batch of queries accumulated
+     * @return null if we have to stop immediately, or a new batch otherwise.
      */
-    static class Rows {
-        List<Row> allRows;
-
-        Rows() {
-            this.allRows = new ArrayList<>();
-        }
-
-        void add(Row row) {
-            this.allRows.add(row);
-        }
-
-        @Override public String toString() {
-            return String.join(System.lineSeparator(),
-                    Linq.map(this.allRows, Row::toString));
-        }
-
-        /**
-         * @return Number of rows.
-         */
-        public int size() {
-            return this.allRows.size();
-        }
-
-        /**
-         * Sort the rows using the specified sort order.
-         */
-        public void sort(SqlTestQueryOutputDescription.SortOrder order) {
-            switch (order) {
-                case NONE:
-                    break;
-                case ROW:
-                    this.allRows.sort(new RowComparator());
-                    break;
-                case VALUE:
-                    this.allRows = net.hydromatic.sqllogictest.util.Utilities.flatMap(this.allRows,
-                            r -> Linq.map(r.values,
-                                    r0 -> {
-                                        Row res = new Row();
-                                        res.add(r0);
-                                        return res;
-                                    }));
-                    this.allRows.sort(new RowComparator());
-                    break;
-            }
-        }
-    }
-
-    Row getValue(DBSPTupleExpression rs, String columnTypes) {
-        Row row = new Row();
-        for(int i = 0; i < columnTypes.length(); ++i) {
-            char c = columnTypes.charAt(i);
-            DBSPLiteral value = rs.get(i).to(DBSPLiteral.class);
-            if (value.isNull) {
-                row.add("NULL");
-                continue;
-            }
-            switch (c) {
-                case 'I': {
-                    long integer;
-                    switch (value.getType().code) {
-                        case BOOL:
-                            integer = Objects.requireNonNull(value.to(DBSPBoolLiteral.class).value) ? 1 : 0;
-                            break;
-                        case DECIMAL:
-                            integer = Objects.requireNonNull(value.to(DBSPDecimalLiteral.class).value).longValue();
-                            break;
-                        case DOUBLE:
-                            integer = Objects.requireNonNull(value.to(DBSPDoubleLiteral.class).value).longValue();
-                            break;
-                        case FLOAT:
-                            integer = Objects.requireNonNull(value.to(DBSPFloatLiteral.class).value).longValue();
-                            break;
-                        case INT32:
-                            integer = Objects.requireNonNull(value.to(DBSPI32Literal.class).value);
-                            break;
-                        case INT64:
-                            integer = Objects.requireNonNull(value.to(DBSPI64Literal.class).value);
-                            break;
-                        case STRING:
-                            try {
-                                integer = Long.parseLong(Objects.requireNonNull(value.to(DBSPStringLiteral.class).value));
-                            } catch (NumberFormatException ex) {
-                                integer = 0;
-                            }
-                            break;
-                        default:
-                            integer = 0;
-                            break;
-                    }
-                    row.add(String.format("%d", integer));
-                    break;
-                }
-                case 'R': {
-                    double d;
-                    switch (value.getType().code) {
-                        case BOOL:
-                            d = Objects.requireNonNull(value.to(DBSPBoolLiteral.class).value) ? 1 : 0;
-                            break;
-                        case DECIMAL:
-                            d = Objects.requireNonNull(value.to(DBSPDecimalLiteral.class).value).doubleValue();
-                            break;
-                        case DOUBLE:
-                            d = Objects.requireNonNull(value.to(DBSPDoubleLiteral.class).value);
-                            break;
-                        case FLOAT:
-                            d = Objects.requireNonNull(value.to(DBSPFloatLiteral.class).value);
-                            break;
-                        case INT32:
-                            d = Objects.requireNonNull(value.to(DBSPI32Literal.class).value);
-                            break;
-                        case INT64:
-                            d = Objects.requireNonNull(value.to(DBSPI64Literal.class).value);
-                            break;
-                        case STRING:
-                            try {
-                                d = Double.parseDouble(Objects.requireNonNull(value.to(DBSPStringLiteral.class).value));
-                            } catch (NumberFormatException ex) {
-                                d = 0;
-                            }
-                            break;
-                        default:
-                            d = 0;
-                            break;
-                    }
-                    row.add(String.format("%.3f", d));
-                    break;
-                }
-                case 'T': {
-                    String s;
-                    switch (value.getType().code) {
-                        case BOOL:
-                            s = Objects.requireNonNull(value.to(DBSPBoolLiteral.class).value) ? "1" : "0";
-                            break;
-                        case DECIMAL:
-                            s = Objects.requireNonNull(value.to(DBSPDecimalLiteral.class).value).toString();
-                            break;
-                        case DOUBLE:
-                            s = Objects.requireNonNull(value.to(DBSPDoubleLiteral.class).value).toString();
-                            break;
-                        case FLOAT:
-                            s = Objects.requireNonNull(value.to(DBSPFloatLiteral.class).value).toString();
-                            break;
-                        case INT32:
-                            s = Objects.requireNonNull(value.to(DBSPI32Literal.class).value).toString();
-                            break;
-                        case INT64:
-                            s = Objects.requireNonNull(value.to(DBSPI64Literal.class).value).toString();
-                            break;
-                        case STRING:
-                            s = Objects.requireNonNull(value.to(DBSPStringLiteral.class).value);
-                            break;
-                        default:
-                            throw new RuntimeException("Unexpected type " + value);
-                    }
-                    StringBuilder result = new StringBuilder();
-
-                    for (int j = 0; j < s.length(); ++j) {
-                        char sc = s.charAt(j);
-                        if (sc < ' ' || sc > '~') {
-                            sc = '@';
-                        }
-
-                        result.append(sc);
-                    }
-
-                    row.add(result.toString());
-                    break;
-                }
-                default:
-                    throw new RuntimeException("Unexpected column type " + c);
-            }
-        }
-        return row;
-    }
-
-    /**
-     * Validate output.  Return 'true' if we need to stop executing.
-     * Reports errors on validation failures.
-     */
-    boolean validateOutput(SqlTestQuery query, int queryNo,
-                           DBSPZSetLiteral.Contents actual,
-                           SqlTestQueryOutputDescription description,
-                           TestStatistics statistics) throws NoSuchAlgorithmException {
-        Rows rows = new Rows();
-        for (Map.Entry<DBSPExpression, Long> entry: actual.data.entrySet()) {
-            if (entry.getValue() <= 0)
-                return statistics.addFailure(
-                        new TestStatistics.FailedTestDescription(query,
-                                "Produced negative result", entry.toString(), null));
-            for (long i = 0; i < entry.getValue(); i++) {
-                DBSPTupleExpression tuple = entry.getKey().to(DBSPTupleExpression.class);
-                Row row = this.getValue(tuple, Objects.requireNonNull(description.columnTypes));
-                rows.add(row);
-            }
-        }
-        if (description.getValueCount()
-                != rows.size() * Objects.requireNonNull(description.columnTypes).length()) {
-            return statistics.addFailure(
-                    new TestStatistics.FailedTestDescription(query,
-                            "Expected " + description.getValueCount() + " values, got "
-                                    + rows.size() * description.columnTypes.length(),
-                            "",
-                            null));
-        }
-        rows.sort(description.getOrder());
-        if (description.getQueryResults() != null) {
-            String r = rows.toString();
-            String q = String.join(System.lineSeparator(),
-                    description.getQueryResults());
-            if (!r.equals(q)) {
-                return statistics.addFailure(
-                        new TestStatistics.FailedTestDescription(query,
-                                "#" + queryNo + " Output differs from expected value",
-                                "computed" + System.lineSeparator()
-                                        + r + System.lineSeparator()
-                                        + "Expected:" + System.lineSeparator()
-                                        + q + System.lineSeparator(),
-                                null));
-            }
-        }
-        if (description.hash != null) {
-            // MD5 is considered insecure, but we have no choice because this is
-            // the algorithm used to compute the checksums by SLT.
-            MessageDigest md = MessageDigest.getInstance("MD5");
-            String repr = rows + System.lineSeparator();
-            md.update(repr.getBytes(StandardCharsets.UTF_8));
-            byte[] digest = md.digest();
-            String hash = net.hydromatic.sqllogictest.util.Utilities.toHex(digest);
-            if (!description.hash.equals(hash)) {
-                return statistics.addFailure(
-                        new TestStatistics.FailedTestDescription(query,
-                                "Hash of data does not match expected value",
-                                "expected:" + description.hash + " "
-                                        + "computed: " + hash + System.lineSeparator(), null));
-            }
-        }
-        return false;
+    @Nullable
+    JitTestBatch runBatch(JitTestBatch batch, TestStatistics result) throws SQLException, NoSuchAlgorithmException {
+        System.out.println("Running a batch of " + batch.size() + " queries");
+        batch.prepareInputs(this.tablePreparation.statements);
+        batch.setInputContents(this.getInputSets());
+        boolean failed = batch.run(result);
+        if (failed)
+            return null;
+        return batch.nextBatch();
     }
 
     @Override
@@ -524,46 +129,61 @@ public class JitDbspExecutor extends SqlSltTestExecutor {
                 options.stopAtFirstError, options.verbosity);
         result.incFiles();
         int queryNo = 0;
-        int skip = 0;  // used only for debugging
-        for (ISqlTestOperation operation : testFile.fileContents) {
+        int batchSize = 1;
+        int skip = 4428;  // used only for debugging
+
+        int remainingInBatch = batchSize;
+        JitTestBatch batch = new JitTestBatch(
+                this.compilerOptions, this.options, rustDirectory, jitExecutableDirectory, queryNo);
+        for (ISqlTestOperation operation: testFile.fileContents) {
             SltSqlStatement stat = operation.as(SltSqlStatement.class);
             if (stat != null) {
+                if (batch.hasQueries()) {
+                    batch = this.runBatch(batch, result);
+                    if (batch == null)
+                        return result;
+                    remainingInBatch = batchSize;
+                }
                 try {
-                    this.statement(stat);
-                    if (!stat.shouldPass) {
-                        options.err.println("Statement should have failed: " + operation);
+                    if (this.buggyOperations.contains(stat.statement)) {
+                        this.options.message("Skipping buggy test " + stat.statement + "\n", 1);
+                    } else {
+                        this.statement(stat);
+                        if (!stat.shouldPass) {
+                            options.err.println("Statement should have failed: " + operation);
+                        }
                     }
                 } catch (SQLException ex) {
-                    // errors in statements cannot be recovered.
-                    if (stat.shouldPass) {
-                        // shouldPass should always be true, otherwise
-                        // the exception should not be thrown.
-                        options.err.println("Error '" + ex.getMessage()
-                                + "' in SQL statement " + operation);
-                        result.incFilesNotParsed();
-                        return result;
-                    }
+                    if (stat.shouldPass)
+                        this.options.error(ex);
                 }
+                this.statementsExecuted++;
             } else {
                 SqlTestQuery query = operation.to(options.err, SqlTestQuery.class);
-                boolean stop = false;
-                try {
-                    if (queryNo >= skip) {
-                        stop = this.query(query, result, queryNo);
-                    }
-                    queryNo++;
-                } catch (Throwable ex) {
-                    // Need to catch Throwable to handle assertion failures too
-                    options.message("Exception during query: " + query.getQuery() + queryNo +
-                            ": " + ex.getMessage(), 1);
-                    stop = result.addFailure(
-                            new TestStatistics.FailedTestDescription(query,
-                                    null, "", ex));
+                if (this.buggyOperations.contains(query.getQuery())) {
+                    options.message("Skipping " + query.getQuery(), 2);
+                    result.incIgnored();
+                    continue;
                 }
-                if (stop) {
-                    break;
+                if (skip > 0) {
+                    skip--;
+                    continue;
+                }
+                // Debugging code commented-out.
+                // if (!"label-510".equals(query.getName())) continue;
+                batch.addQuery(query);
+                remainingInBatch--;
+                queryNo++;
+                if (remainingInBatch == 0) {
+                    batch = this.runBatch(batch, result);
+                    if (batch == null)
+                        return result;
+                    remainingInBatch = batchSize;
                 }
             }
+        }
+        if (remainingInBatch != batchSize) {
+            this.runBatch(batch, result);
         }
         this.statementExecutor.dropAllViews();
         this.statementExecutor.dropAllTables();
@@ -582,6 +202,7 @@ public class JitDbspExecutor extends SqlSltTestExecutor {
                 compilerOptions.ioOptions.jit = true;
                 compilerOptions.optimizerOptions.throwOnError = options.stopAtFirstError;
                 compilerOptions.ioOptions.lenient = true;
+                compilerOptions.ioOptions.quiet = true;
                 JitDbspExecutor result = new JitDbspExecutor(
                         Objects.requireNonNull(inner), options, compilerOptions);
                 Set<String> bugs = options.readBugsFile();
